modify インスタンスの復元を factory -> repository に移動
